requirements/dev.txt
-r base.txt

coverage
ipython
pycodestyle<2.4
flake8<3.6
tox
mock
factory-boy
swapper
mysqlclient
django-polymorphic
